+++ internal/pkg/job/model.go
@@ -1,34 +1,17 @@
 package job
 
 import (
-	"io/ioutil"
-	"time"
-
 	"gopkg.in/yaml.v2"
+	"io/ioutil"
+	"net/http"
+	"reflect"
 )
 
 // Descriptor represents a JOB document
 type Descriptor struct {
-	Name    string            `json:"project,omitempty" yaml:"project,omitempty"`
-	UpJobs  map[string]*UpJob `json:"up_jobs,omitempty" yaml:"up_jobs,omitempty"`
-	DNSJobs map[string]DnsJob `json:"dns_jobs,omitempty" yaml:"dns_jobs,omitempty"`
-}
-
-func DescriptorFromFile(path string) (Descriptor, error) {
-	d := Descriptor{}
-	data, _ := ioutil.ReadFile(path)
-	err := yaml.Unmarshal(data, &d)
-
-	for name, upJob := range d.UpJobs {
-		upJob.Name = name
-		d.UpJobs[name] = upJob
-	}
-
-	for name, dnsJob := range d.DNSJobs {
-		dnsJob.Name = name
-		d.DNSJobs[name] = dnsJob
-	}
-	return d, err
+	Name    string             `json:"project,omitempty" yaml:"project,omitempty"`
+	UpJobs  map[string]*UpJob  `json:"up_jobs,omitempty" yaml:"up_jobs,omitempty"`
+	DNSJobs map[string]*DnsJob `json:"dns_jobs,omitempty" yaml:"dns_jobs,omitempty"`
 }
 
 // UpJob is a check if a HTTP endpoint is up and able to serve required method
@@ -36,7 +19,7 @@ type UpJob struct {
 	Name                string
 	Host                string            `json:"host" yaml:"host"`
 	URL                 string            `json:"url" yaml:"url"`
-	Method              Method            `json:"method,omitempty" yaml:"method,omitempty"`
+	Method              string            `json:"method,omitempty" yaml:"method,omitempty"`
 	Expected            int               `json:"expected_code,omitempty" yaml:"expected_code,omitempty"`
 	Headers             map[string]string `json:"headers,omitempty" yaml:"headers,omitempty"`
 	PlainBody           string            `json:"plain_body,omitempty" yaml:"plain_body,omitempty"`
@@ -46,63 +29,53 @@ type UpJob struct {
 	ContentMatch        string            `json:"content_match,omitempty" yaml:"content_match,omitempty"`
 	ReverseContentMatch bool              `json:"reverse_content_match,omitempty" yaml:"reverse_content_match,omitempty"`
 	Oauth               Oauth             `json:"Oauth,omitempty" yaml:"Oauth,omitempty"`
-	OauthServerResponse OauthResponse
+	Context             Context
+	RequestInterceptors []HttpRequestInterceptor
 }
 
-type Oauth struct {
-	AuthUrl string            `json:"auth_url,omitempty" yaml:"auth_url,omitempty"`
-	Params  map[string]string `json:"params,omitempty" yaml:"params,omitempty"`
-	Headers map[string]string `json:"headers,omitempty" yaml:"headers,omitempty"`
+func DescriptorFromFile(path string) (Descriptor, error) {
+	d := Descriptor{}
+	data, _ := ioutil.ReadFile(path)
+	err := yaml.Unmarshal(data, &d)
+
+	for name, upJob := range d.UpJobs {
+		upJob.Name = name
+		//Check, if upJob implemets Job interface, and call init(), in case it does
+		value := reflect.ValueOf(upJob)
+		if job, yes := value.Interface().(Job); yes {
+			job.init()
+		}
+	}
+
+	for name, dnsJob := range d.DNSJobs {
+		dnsJob.Name = name
+	}
+	return d, err
 }
 
-type OauthRequest struct {
-	GrantType    string `json:"grant_type,omitempty" yaml:"grant_type,omitempty"`
-	ClientId     string `json:"client_id,omitempty" yaml:"client_id,omitempty"`
-	ClientSecret int    `json:"client_secret,omitempty" yaml:"client_secret,omitempty"`
-	Username     string `json:"username,omitempty" yaml:"username,omitempty"`
-	Password     string `json:"password,omitempty" yaml:"password,omitempty"`
+type Job interface {
+	init()
 }
 
-type OauthResponse struct {
-	AccessToken  string `json:"access_token,omitempty" yaml:"access_token,omitempty"`
-	TokenType    string `json:"token_type,omitempty" yaml:"token_type,omitempty"`
-	ExpiresIn    int    `json:"expires_in,omitempty" yaml:"expires_in,omitempty"`
-	RefreshToken string `json:"refresh_token,omitempty" yaml:"refresh_token,omitempty"`
-	Scope        string `json:"scope,omitempty" yaml:"scope,omitempty"`
-	ExpiresAt    time.Time
-	Refresh      bool
+type Context struct {
+	OauthResponse OauthResponse
 }
-
-type UpJobs map[string]UpJob
 
 //Defining default values for unmarshalling UpJob
 type tmpUpJob UpJob
 
-func (j *UpJob) UnmarshalYAML(unmarshal func(interface{}) error) error {
+func (job *UpJob) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	tmp := &tmpUpJob{
-		Method:              GET,
+		Method:              http.MethodGet,
 		Expected:            200,
 		ContentMatch:        "",
 		ReverseContentMatch: false,
 		CheckSSL:            true,
 	}
 	unmarshal(tmp)
-	*j = UpJob(*tmp)
+	*job = UpJob(*tmp)
 	return nil
 }
-
-// Method is a HTTP method
-type Method string
-
-const (
-	GET     Method = "GET"
-	POST    Method = "POST"
-	PUT     Method = "PUT"
-	DELETE  Method = "DELETE"
-	PATCH   Method = "PATCH"
-	OPTIONS Method = "OPTIONS"
-	HEAD    Method = "HEAD"
-)
 
 // DnsJob is verifying if a fqdn is looked up to the expected set of IPs
 type DnsJob struct {
@@ -111,4 +84,8 @@ type DnsJob struct {
 	FQDN       string            `json:"fqdn" yaml:"fqdn"`
 	IPs        []string          `json:"ips" yaml:"ips"`
 	CustomTags map[string]string `json:"tags,omitempty" yaml:"tags,omitempty"`
+}
+
+type HttpRequestInterceptor interface {
+	Intercept(req *http.Request, upJob *UpJob) *http.Request
 }