+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- Introduced extension mechanism via HttpRequestInterceptor interface
+- Oauth Authorization with grant_type `client_credentials` and `password`
 
 ### Changed